% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_format_mapping}
\alias{apply_format_mapping}
\title{Apply format mapping to data}
\usage{
apply_format_mapping(data, format_df)
}
\arguments{
\item{data}{Data frame or list of data frames containing study data}

\item{format_df}{Data frame containing format mapping information with columns:
FMTNAME, START, and LABEL}
}
\value{
Data frame or list of data frames with values mapped according to format_df
}
\description{
Map values in data frames based on format information, similar to SAS formats.
This function applies value mappings defined in a format data frame to corresponding
columns in the input data.
}
\concept{format utilities}
