% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age_requirement.R
\name{check_age_requirement}
\alias{check_age_requirement}
\title{Check for patients under 18 years old}
\usage{
check_age_requirement(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check for patients under 18 years old
}
