% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_check_ascvd_ldl_random.R
\name{check_ascvd_ldl_random}
\alias{check_ascvd_ldl_random}
\title{Check ASCVD history and LDL-C random criteria}
\usage{
check_ascvd_ldl_random(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check ASCVD history and LDL-C random criteria
}
