% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ck.R
\name{check_ck_exclusion}
\alias{check_ck_exclusion}
\title{Check CK level exclusion criteria}
\usage{
check_ck_exclusion(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check CK level exclusion criteria
}
