% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dose_site_requirement.R
\name{check_dose_site_requirement}
\alias{check_dose_site_requirement}
\title{Check dose and injection site requirements}
\usage{
check_dose_site_requirement(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check dose and injection site requirements
}
