% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gfr_requirement.R
\name{check_gfr_requirement}
\alias{check_gfr_requirement}
\title{Check eGFR requirements}
\usage{
check_gfr_requirement(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check eGFR requirements
}
