% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ldl_requirement.R
\name{check_ldl_requirement}
\alias{check_ldl_requirement}
\title{Check LDL-C requirements}
\usage{
check_ldl_requirement(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check LDL-C requirements
}
