% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_liver.R
\name{check_liver_exclusion}
\alias{check_liver_exclusion}
\title{Check liver function test exclusion criteria}
\usage{
check_liver_exclusion(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check liver function test exclusion criteria
}
