% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_medication_window.R
\name{check_medication_window}
\alias{check_medication_window}
\title{Check study medication administration time window}
\usage{
check_medication_window(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check study medication administration time window
}
