% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_history.R
\name{check_missing_history}
\alias{check_missing_history}
\title{Check for missing medical history data}
\usage{
check_missing_history(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List of data frames with missing history information
}
\description{
Check for missing medical history data
}
