% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_visit.R
\name{check_missing_visit}
\alias{check_missing_visit}
\title{Check missing visits}
\usage{
check_missing_visit(test_data, dataset_name)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check missing visits
}
