% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nyha.R
\name{check_nyha_requirement}
\alias{check_nyha_requirement}
\title{Check NYHA class exclusion criteria}
\usage{
check_nyha_requirement(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check NYHA class exclusion criteria
}
