% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pregnancy.R
\name{check_pregnancy_exclusion}
\alias{check_pregnancy_exclusion}
\title{Check pregnancy test exclusion criteria}
\usage{
check_pregnancy_exclusion(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check pregnancy test exclusion criteria
}
