% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_strata.R
\name{check_random_deviation}
\alias{check_random_deviation}
\title{Check randomization stratification factor deviation}
\usage{
check_random_deviation(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check randomization stratification factor deviation
}
