% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_statin_random.R
\name{check_statin_random}
\alias{check_statin_random}
\title{Check prior statin therapy random criteria}
\usage{
check_statin_random(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check prior statin therapy random criteria
}
