% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_check_tg_requirement.R
\name{check_tg_requirement}
\alias{check_tg_requirement}
\title{Check triglyceride (TG) requirements}
\usage{
check_tg_requirement(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check triglyceride (TG) requirements
}
