% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_thyroid.R
\name{check_thyroid_exclusion}
\alias{check_thyroid_exclusion}
\title{Check thyroid function exclusion criteria}
\usage{
check_thyroid_exclusion(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check thyroid function exclusion criteria
}
