% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_virus.R
\name{check_virus_exclusion}
\alias{check_virus_exclusion}
\title{Check viral infection exclusion criteria}
\usage{
check_virus_exclusion(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check viral infection exclusion criteria
}
