% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_check_results}
\alias{combine_check_results}
\title{Combine multiple check results into a single data frame}
\usage{
combine_check_results(...)
}
\arguments{
\item{...}{Multiple check result objects}
}
\value{
Data frame with combined check results
}
\description{
Combine multiple check results into a single data frame
}
