% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pdchecker_options}
\alias{get_pdchecker_options}
\title{Get pdchecker Global Options}
\usage{
get_pdchecker_options()
}
\value{
A named list of current pdchecker option values.
Options that have not been set will show their default values.
}
\description{
Display the current values of all pdchecker global options.
}
\examples{
\dontrun{
# View current options
get_pdchecker_options()

# Set some options and verify
set_pdchecker_options(sv_dataset = "SV", ex_datasets = c("EC"))
get_pdchecker_options()
}

}
\seealso{
\code{\link[=set_pdchecker_options]{set_pdchecker_options()}} for setting option values

Other options: 
\code{\link{set_pdchecker_options}()}
}
\concept{options}
