% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age_requirement.R
\name{print.age_check}
\alias{print.age_check}
\title{Print method for age requirement check results}
\usage{
\method{print}{age_check}(x, ...)
}
\arguments{
\item{x}{Object of class age_check}

\item{...}{Additional arguments}
}
\description{
Print method for age requirement check results
}
