% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ascvd_ldl_random.R
\name{print.ascvd_ldl_check}
\alias{print.ascvd_ldl_check}
\title{Print method for ASCVD and LDL-C check results}
\usage{
\method{print}{ascvd_ldl_check}(x, ...)
}
\arguments{
\item{x}{Object of class ascvd_ldl_check}

\item{...}{Additional arguments}
}
\description{
Print method for ASCVD and LDL-C check results
}
