% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_check_ck.R
\name{print.ck_check}
\alias{print.ck_check}
\title{Print method for CK check results}
\usage{
\method{print}{ck_check}(x, ...)
}
\arguments{
\item{x}{Object of class ck_check}

\item{...}{Additional arguments}
}
\description{
Print method for CK check results
}
