% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dose_site_requirement.R
\name{print.dose_site_check}
\alias{print.dose_site_check}
\title{Print method for dose and site check results}
\usage{
\method{print}{dose_site_check}(x, ...)
}
\arguments{
\item{x}{Object of class dose_site_check}

\item{...}{Additional arguments}
}
\description{
Print method for dose and site check results
}
