% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gfr_requirement.R
\name{print.gfr_check}
\alias{print.gfr_check}
\title{Print method for eGFR requirement check results}
\usage{
\method{print}{gfr_check}(x, ...)
}
\arguments{
\item{x}{Object of class gfr_check}

\item{...}{Additional arguments}
}
\description{
Print method for eGFR requirement check results
}
