% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_icf_time_deviation.R
\name{print.icf_time_deviation}
\alias{print.icf_time_deviation}
\title{Print method for ICF time deviation check results}
\usage{
\method{print}{icf_time_deviation}(x, ...)
}
\arguments{
\item{x}{Object of class icf_time_deviation}

\item{...}{Additional arguments}
}
\description{
Print method for ICF time deviation check results
}
