% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_check_ldl_requirement.R
\name{print.ldl_check}
\alias{print.ldl_check}
\title{Print method for LDL-C requirement check results}
\usage{
\method{print}{ldl_check}(x, ...)
}
\arguments{
\item{x}{Object of class ldl_check}

\item{...}{Additional arguments}
}
\description{
Print method for LDL-C requirement check results
}
