% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_check_liver.R
\name{print.liver_check}
\alias{print.liver_check}
\title{Print method for liver function check results}
\usage{
\method{print}{liver_check}(x, ...)
}
\arguments{
\item{x}{Object of class liver_check}

\item{...}{Additional arguments}
}
\description{
Print method for liver function check results
}
