% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_medication_window.R
\name{print.medication_window_check}
\alias{print.medication_window_check}
\title{Print method for medication window check results}
\usage{
\method{print}{medication_window_check}(x, ...)
}
\arguments{
\item{x}{Object of class medication_window_check}

\item{...}{Additional arguments}
}
\description{
Print method for medication window check results
}
