% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_history.R
\name{print.missing_history_check}
\alias{print.missing_history_check}
\title{Print method for missing history check}
\usage{
\method{print}{missing_history_check}(x, ...)
}
\arguments{
\item{x}{Object of class missing_history_check}

\item{...}{Additional arguments}
}
\value{
Invisible x
}
\description{
Print method for missing history check
}
