% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_test.R
\name{print.missing_test_check}
\alias{print.missing_test_check}
\title{Print method for missing test check results}
\usage{
\method{print}{missing_test_check}(x, ...)
}
\arguments{
\item{x}{Object of class missing_test_check}

\item{...}{Additional arguments}
}
\description{
Print method for missing test check results
}
