% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_visit.R
\name{print.missing_visit_check}
\alias{print.missing_visit_check}
\title{Print method for missing visit check results}
\usage{
\method{print}{missing_visit_check}(x, ...)
}
\arguments{
\item{x}{Object of class missing_visit_check}

\item{...}{Additional arguments}
}
\description{
Print method for missing visit check results
}
