% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_check_nyha.R
\name{print.nyha_check}
\alias{print.nyha_check}
\title{Print method for NYHA requirement check results}
\usage{
\method{print}{nyha_check}(x, ...)
}
\arguments{
\item{x}{Object of class nyha_check}

\item{...}{Additional arguments}
}
\description{
Print method for NYHA requirement check results
}
