% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_check_pregnancy.R
\name{print.pregnancy_check}
\alias{print.pregnancy_check}
\title{Print method for pregnancy test check results}
\usage{
\method{print}{pregnancy_check}(x, ...)
}
\arguments{
\item{x}{Object of class pregnancy_check}

\item{...}{Additional arguments}
}
\description{
Print method for pregnancy test check results
}
