% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_check_strata.R
\name{print.rand_deviation}
\alias{print.rand_deviation}
\title{Print method for randomization deviation check results}
\usage{
\method{print}{rand_deviation}(x, ...)
}
\arguments{
\item{x}{Object of class rand_deviation}

\item{...}{Additional arguments}
}
\description{
Print method for randomization deviation check results
}
