% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_screen_without_ic.R
\name{print.screen_ic_check}
\alias{print.screen_ic_check}
\title{Print method for screen without IC check results}
\usage{
\method{print}{screen_ic_check}(x, ...)
}
\arguments{
\item{x}{Object of class screen_ic_check}

\item{...}{Additional arguments}
}
\description{
Print method for screen without IC check results
}
