% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_statin_random.R
\name{print.statin_check}
\alias{print.statin_check}
\title{Print method for statin therapy check results}
\usage{
\method{print}{statin_check}(x, ...)
}
\arguments{
\item{x}{Object of class statin_check}

\item{...}{Additional arguments}
}
\description{
Print method for statin therapy check results
}
