% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tg_requirement.R
\name{print.tg_check}
\alias{print.tg_check}
\title{Print method for TG requirement check results}
\usage{
\method{print}{tg_check}(x, ...)
}
\arguments{
\item{x}{Object of class tg_check}

\item{...}{Additional arguments}
}
\description{
Print method for TG requirement check results
}
