% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_thyroid.R
\name{print.thyroid_check}
\alias{print.thyroid_check}
\title{Print method for thyroid function check results}
\usage{
\method{print}{thyroid_check}(x, ...)
}
\arguments{
\item{x}{Object of class thyroid_check}

\item{...}{Additional arguments}
}
\description{
Print method for thyroid function check results
}
