% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_check_virus.R
\name{print.virus_check}
\alias{print.virus_check}
\title{Print method for viral infection check results}
\usage{
\method{print}{virus_check}(x, ...)
}
\arguments{
\item{x}{Object of class virus_check}

\item{...}{Additional arguments}
}
\description{
Print method for viral infection check results
}
