% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_visit_window.R
\name{print.visit_window_check}
\alias{print.visit_window_check}
\title{Print method for visit window check results}
\usage{
\method{print}{visit_window_check}(x, ...)
}
\arguments{
\item{x}{Object of class visit_window_check}

\item{...}{Additional arguments}
}
\description{
Print method for visit window check results
}
