% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Read raw data}
\usage{
read_raw_data(folder, iwrs_file = "iwrs.csv", format_file = NULL)
}
\arguments{
\item{folder}{Path to data directory}
}
\value{
A named list of data.frames where names are filenames
}
\description{
Scan specified directory and read all SAS format files
}
