% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_format_mapping}
\alias{apply_format_mapping}
\title{Apply format mapping to data}
\usage{
apply_format_mapping(data, format_df)
}
\arguments{
\item{data}{List of data frames containing study data}

\item{format_df}{Data frame containing format information}
}
\value{
List of data frames with mapped values
}
\description{
Map values in data frames based on format information
}
