% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_check_df}
\alias{as_check_df}
\title{Convert check results to a standardized data frame}
\usage{
as_check_df(check_result, check_name = NULL)
}
\arguments{
\item{check_result}{List containing check results from any check function}

\item{check_name}{Name of the check (optional, will be extracted from class if not provided)}
}
\value{
Data frame with standardized check results
}
\description{
Convert check results to a standardized data frame
}
