% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_check_results}
\alias{capture_check_results}
\title{Capture and parse output from multiple check functions}
\usage{
capture_check_results(..., data)
}
\arguments{
\item{...}{Named list of check functions to run}

\item{data}{Data to pass to each check function}
}
\value{
Data frame with combined check results
}
\description{
Capture and parse output from multiple check functions
}
