% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_icf_time_deviation.R
\name{check_icf_time_deviation}
\alias{check_icf_time_deviation}
\title{Check if any time variables occur before informed consent time}
\usage{
check_icf_time_deviation(data, ignore_vars = "BRTHDAT")
}
\arguments{
\item{data}{List of data frames containing study data}

\item{ignore_vars}{Character vector of variables to ignore in the check (default: "BRTHDAT")}
}
\value{
List containing check results and descriptions
}
\description{
Check if any time variables occur before informed consent time
}
