% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_screen_without_ic.R
\name{check_screen_without_ic}
\alias{check_screen_without_ic}
\title{Check for screened subjects without informed consent}
\usage{
check_screen_without_ic(data)
}
\arguments{
\item{data}{List of data frames containing study data}
}
\value{
List containing check results and descriptions
}
\description{
Check for screened subjects without informed consent
}
