% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_test_window.R
\name{print.test_window_check}
\alias{print.test_window_check}
\title{Print method for missing test check results}
\usage{
\method{print}{test_window_check}(x, ...)
}
\arguments{
\item{x}{Object of class missing_test_check}

\item{...}{Additional arguments}
}
\description{
Print method for missing test check results
}
